added raydium cpmm swaps

+++ src/types.rs
@@ -1,6 +1,7 @@
 use solana_sdk::{pubkey::Pubkey, signature::Signature};
+use strum::{Display, EnumIter};
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct Event {
     pub signature: Signature,
     pub event_type: EventType,
@@ -8,7 +9,7 @@ pub struct Event {
     pub timestamp: u64,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum EventType {
     Swap(Swap),
     PoolCreation {
@@ -20,7 +21,7 @@ pub enum EventType {
     },
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct Swap {
     pub token_in_amount: u64,
     pub token_in_decimals: u8,
@@ -32,12 +33,12 @@ pub struct Swap {
 
     pub platform: SwapPlatform,
     /// Including decimals
-    pub token_in_reserve: f64,
+    pub token_in_reserve: u64,
     /// Including decimals
-    pub token_out_reserve: f64,
+    pub token_out_reserve: u64,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Hash, PartialEq, Eq, EnumIter, Display)]
 pub enum SwapPlatform {
     PumpFun,
     Raydium,
